requirements.txt
python-dotenv
fastapi
uvicorn
python-docx
pdfplumber
pytesseract
pdf2image
python-multipart

cryptography

openai
groq
langchain
langchain-openai
langchain-groq
langchain-community
